Tech's suggestions.
