+++ fabric-data-generation-api-v1/src/main/java/net/fabricmc/fabric/api/datagen/v1/provider/FabricLanguageProvider.java
@@ -17,20 +17,33 @@
 package net.fabricmc.fabric.api.datagen.v1.provider;
 
 import java.io.IOException;
+import java.io.Reader;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Map;
+import java.util.Objects;
 import java.util.TreeMap;
 
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import org.jetbrains.annotations.ApiStatus;
 
+import net.minecraft.block.Block;
 import net.minecraft.data.DataProvider;
 import net.minecraft.data.DataWriter;
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.attribute.EntityAttribute;
+import net.minecraft.entity.effect.StatusEffect;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemGroup;
+import net.minecraft.stat.StatType;
+import net.minecraft.util.Identifier;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-import net.fabricmc.fabric.api.datagen.v1.provider.consumers.TranslationConsumer;
 
 /**
- * Extend this class and implement {@link FabricLanguageProvider#generateTranslations(TranslationConsumer)}.
+ * Extend this class and implement {@link FabricLanguageProvider#generateTranslations(TranslationBuilder)}.
  * Make sure to use {@link FabricLanguageProvider#FabricLanguageProvider(FabricDataGenerator, String)} FabricLanguageProvider} to declare what language code is being generated if it isn't {@code en_us}.
  *
  * <p>Register an instance of the class with {@link FabricDataGenerator#addProvider} in a {@link net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint}
@@ -51,15 +64,18 @@ public abstract class FabricLanguageProvider implements DataProvider {
 	/**
 	 * Implement this method to register languages.
 	 *
-	 * <p>Call {@link TranslationConsumer#add(String, String)} to add a translation.
+	 * <p>Call {@link TranslationBuilder#add(String, String)} to add a translation.
 	 */
-	public abstract void generateTranslations(TranslationConsumer translationConsumer);
+	public abstract void generateTranslations(TranslationBuilder translationBuilder);
 
 	@Override
 	public void run(DataWriter writer) throws IOException {
 		TreeMap<String, String> translationEntries = new TreeMap<>();
 
 		generateTranslations((String key, String value) -> {
+			Objects.requireNonNull(key);
+			Objects.requireNonNull(value);
+
 			if (translationEntries.containsKey(key)) {
 				throw new RuntimeException("Existing translation key found - " + key + " - Duplicate will be ignored.");
 			}
@@ -83,5 +99,116 @@ public abstract class FabricLanguageProvider implements DataProvider {
 	@Override
 	public String getName() {
 		return "Language";
+	}
+
+	/**
+	 * A consumer used by {@link FabricLanguageProvider#generateTranslations(TranslationBuilder)}.
+	 */
+	@ApiStatus.NonExtendable
+	@FunctionalInterface
+	public interface TranslationBuilder {
+		/**
+		 * Adds a language entry.
+		 *
+		 * @param languageKey  The key of the language entry.
+		 * @param value        The value of the entry.
+		 */
+		void add(String languageKey, String value);
+
+		/**
+		 * Adds a language entry for an {@link Item}.
+		 * @param item The {@link Item} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(Item item, String value) {
+			add(item.getTranslationKey(), value);
+		};
+
+		/**
+		 * Adds a language entry for a {@link Block}.
+		 * @param block The {@link Block} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(Block block, String value) {
+			add(block.getTranslationKey(), value);
+		}
+
+		/**
+		 * Adds a language entry for an {@link ItemGroup}.
+		 * @param group The {@link ItemGroup} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(ItemGroup group, String value) {
+			add("itemGroup." + group.getName(), value);
+		}
+
+		/**
+		 * Adds a language entry for an {@link EntityType}.
+		 * @param entityType The {@link EntityType} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(EntityType<?> entityType, String value) {
+			add(entityType.getTranslationKey(), value);
+		}
+
+		/**
+		 * Adds a language entry for an {@link Enchantment}.
+		 * @param enchantment The {@link Enchantment} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(Enchantment enchantment, String value) {
+			add(enchantment.getTranslationKey(), value);
+		}
+
+		/**
+		 * Adds a language entry for an {@link EntityAttribute}.
+		 * @param entityAttribute The {@link EntityAttribute} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(EntityAttribute entityAttribute, String value) {
+			add(entityAttribute.getTranslationKey(), value);
+		}
+
+		/**
+		 * Adds a language entry for a {@link StatType}.
+		 * @param statType The {@link StatType} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(StatType<?> statType, String value) {
+			add(statType.getTranslationKey(), value);
+		}
+
+		/**
+		 * Adds a language entry for a {@link StatusEffect}.
+		 * @param statusEffect The {@link StatusEffect} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(StatusEffect statusEffect, String value) {
+			add(statusEffect.getTranslationKey(), value);
+		}
+
+		/**
+		 * Adds a language entry for an {@link Identifier}.
+		 * @param identifier The {@link Identifier} to get the language entry key from.
+		 * @param value The value of the entry.
+		 */
+		default void add(Identifier identifier, String value) {
+			add(identifier.toTranslationKey(), value);
+		}
+
+		/**
+		 * Merges an existing language file into the generated language file.
+		 * @param existingLanguageFile The path to the existing language file.
+		 * @throws IOException If loading the language file failed.
+		 */
+		default void add(Path existingLanguageFile) throws IOException {
+			try (Reader reader = Files.newBufferedReader(existingLanguageFile)) {
+				JsonObject translations = JsonParser.parseReader(reader).getAsJsonObject();
+
+				for (String key : translations.keySet()) {
+					add(key, translations.get(key).getAsString());
+				}
+			}
+		}
 	}
 }
